+++ src/main/java/com/myprecious/moneyglove/domain/board/repository/BoardRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface BoardRepository extends JpaRepository<BoardEntity, Long> {
     List<BoardEntity> findByUser_Uid(String uid);
     List<BoardEntity> findByBoardStatus(BoardEntity.BoardStatus boardStatus);
-    List<BoardEntity> findBydDay(Long dDay);
+    List<BoardEntity> findBydDay(Integer dDay);
 }
